Cleaning up unused members.
Added XML comments
