+++ Moq.AutoMock/AutoMocker.Verify.cs
@@ -5,7 +5,12 @@ namespace Moq.AutoMock
 {
     public partial class AutoMocker
     {
-
+        /// <summary>
+        /// Verifies that a specific invocation matching the given expression was performed
+        /// on the mock. Use in conjunction with the default Moq.MockBehavior.Loose.
+        /// </summary>
+        /// <typeparam name="T">Type of the mock</typeparam>
+        /// <param name="expression">Expression to verify</param>
         public void Verify<T>(Expression<Action<T>> expression)
             where T : class
         {
@@ -14,6 +19,13 @@ namespace Moq.AutoMock
         }
 
 
+        /// <summary>
+        /// Verifies that a specific invocation matching the given expression was performed
+        /// on the mock. Use in conjunction with the default Moq.MockBehavior.Loose.
+        /// </summary>
+        /// <typeparam name="T">Type of the mock</typeparam>
+        /// <param name="expression">Expression to verify</param>
+        /// <param name="times">The number of times a method is allowed to be called.</param>
         public void Verify<T>(Expression<Action<T>> expression, Times times)
             where T : class
         {
@@ -21,7 +33,13 @@ namespace Moq.AutoMock
             mock.Verify(expression, times);
         }
 
-
+        /// <summary>
+        /// Verifies that a specific invocation matching the given expression was performed
+        /// on the mock. Use in conjunction with the default Moq.MockBehavior.Loose.
+        /// </summary>
+        /// <typeparam name="T">Type of the mock</typeparam>
+        /// <param name="expression">Expression to verify</param>
+        /// <param name="times">The number of times a method is allowed to be called.</param>
         public void Verify<T>(Expression<Action<T>> expression, Func<Times> times)
             where T : class
         {
@@ -29,31 +47,56 @@ namespace Moq.AutoMock
             mock.Verify(expression, times);
         }
 
-
-        public void Verify<T>(Expression<Action<T>> expression, String failMessage)
+        /// <summary>
+        /// Verifies that a specific invocation matching the given expression was performed
+        /// on the mock. Use in conjunction with the default Moq.MockBehavior.Loose.
+        /// </summary>
+        /// <typeparam name="T">Type of the mock</typeparam>
+        /// <param name="expression">Expression to verify</param>
+        /// <param name="failMessage">Message to show if verification fails.</param>
+        public void Verify<T>(Expression<Action<T>> expression, string failMessage)
             where T : class
         {
             var mock = GetMock<T>();
             mock.Verify(expression, failMessage);
         }
 
-
-        public void Verify<T>(Expression<Action<T>> expression, Times times, String failMessage)
+        /// <summary>
+        /// Verifies that a specific invocation matching the given expression was performed
+        /// on the mock. Use in conjunction with the default Moq.MockBehavior.Loose.
+        /// </summary>
+        /// <typeparam name="T">Type of the mock</typeparam>
+        /// <param name="expression">Expression to verify</param>
+        /// <param name="times">The number of times a method is allowed to be called.</param>
+        /// <param name="failMessage">Message to show if verification fails.</param>
+        public void Verify<T>(Expression<Action<T>> expression, Times times, string failMessage)
             where T : class
         {
             var mock = GetMock<T>();
             mock.Verify(expression, times, failMessage);
         }
 
-
-        public void Verify<T>(Expression<Action<T>> expression, Func<Times> times, String failMessage)
+        /// <summary>
+        /// Verifies that a specific invocation matching the given expression was performed
+        /// on the mock. Use in conjunction with the default Moq.MockBehavior.Loose.
+        /// </summary>
+        /// <typeparam name="T">Type of the mock</typeparam>
+        /// <param name="expression">Expression to verify</param>
+        /// <param name="times">The number of times a method is allowed to be called.</param>
+        /// <param name="failMessage">Message to show if verification fails.</param>
+        public void Verify<T>(Expression<Action<T>> expression, Func<Times> times, string failMessage)
             where T : class
         {
             var mock = GetMock<T>();
             mock.Verify(expression, times, failMessage);
         }
 
-
+        /// <summary>
+        /// Verifies that a specific invocation matching the given expression was performed
+        /// on the mock. Use in conjunction with the default Moq.MockBehavior.Loose.
+        /// </summary>
+        /// <typeparam name="T">Type of the mock</typeparam>
+        /// <param name="expression">Expression to verify</param>
         public void Verify<T>(Expression<Func<T, object>> expression)
             where T : class
         {
@@ -65,7 +108,13 @@ namespace Moq.AutoMock
             mock.Verify(expression);
         }
 
-
+        /// <summary>
+        /// Verifies that a specific invocation matching the given expression was performed
+        /// on the mock. Use in conjunction with the default Moq.MockBehavior.Loose.
+        /// </summary>
+        /// <typeparam name="T">Type of the mock</typeparam>
+        /// <param name="expression">Expression to verify</param>
+        /// <param name="times">The number of times a method is allowed to be called.</param>
         public void Verify<T>(Expression<Func<T, object>> expression, Times times)
             where T : class
         {
@@ -77,7 +126,13 @@ namespace Moq.AutoMock
             mock.Verify(expression, times);
         }
 
-
+        /// <summary>
+        /// Verifies that a specific invocation matching the given expression was performed
+        /// on the mock. Use in conjunction with the default Moq.MockBehavior.Loose.
+        /// </summary>
+        /// <typeparam name="T">Type of the mock</typeparam>
+        /// <param name="expression">Expression to verify</param>
+        /// <param name="times">The number of times a method is allowed to be called.</param>
         public void Verify<T>(Expression<Func<T, object>> expression, Func<Times> times)
             where T : class
         {
@@ -89,7 +144,14 @@ namespace Moq.AutoMock
             mock.Verify(expression, times);
         }
 
-        public void Verify<T>(Expression<Func<T, object>> expression, String failMessage)
+        /// <summary>
+        /// Verifies that a specific invocation matching the given expression was performed
+        /// on the mock. Use in conjunction with the default Moq.MockBehavior.Loose.
+        /// </summary>
+        /// <typeparam name="T">Type of the mock</typeparam>
+        /// <param name="expression">Expression to verify</param>
+        /// <param name="failMessage">Message to show if verification fails.</param>
+        public void Verify<T>(Expression<Func<T, object>> expression, string failMessage)
             where T : class
         {
             var mock = GetMock<T>();
@@ -100,8 +162,15 @@ namespace Moq.AutoMock
             mock.Verify(expression, failMessage);
         }
 
-
-        public void Verify<T>(Expression<Func<T, object>> expression, Times times, String failMessage)
+        /// <summary>
+        /// Verifies that a specific invocation matching the given expression was performed
+        /// on the mock. Use in conjunction with the default Moq.MockBehavior.Loose.
+        /// </summary>
+        /// <typeparam name="T">Type of the mock</typeparam>
+        /// <param name="expression">Expression to verify</param>
+        /// <param name="times">The number of times a method is allowed to be called.</param>
+        /// <param name="failMessage">Message to show if verification fails.</param>
+        public void Verify<T>(Expression<Func<T, object>> expression, Times times, string failMessage)
             where T : class
         {
             var mock = GetMock<T>();
@@ -112,7 +181,11 @@ namespace Moq.AutoMock
             mock.Verify(expression, times, failMessage);
         }
 
-
+        /// <summary>
+        /// Verifies that a specific invocation matching the given expression was performed
+        /// on the mock. Use in conjunction with the default Moq.MockBehavior.Loose.
+        /// </summary>
+        /// <typeparam name="T">Type of the mock</typeparam>
         public void Verify<T>()
             where T : class
         {
