+++ Moq.AutoMock/Moq.AutoMock.csproj
@@ -7,19 +7,17 @@
     <AssemblyName>Moq.AutoMock</AssemblyName>
     <AssemblyTitle>Moq.AutoMocker</AssemblyTitle>
     <Product>Moq.AutoMocker</Product>
-    <Copyright>Copyright © 2018</Copyright>
+    <Copyright>Copyright © 2020</Copyright>
 
     <Version>0.5.0</Version>
     <Authors>Tim Kellogg,Adam Hewitt,Kevin Bost</Authors>
     <PackageDescription>An auto-mocking container that generates mocks using Moq</PackageDescription>
-    <Copyright>Copyright Tim Kellogg 2018</Copyright>
+    <Copyright>Copyright Tim Kellogg 2020</Copyright>
     <PackageLicenseFile>LICENSE</PackageLicenseFile>
     <PackageProjectUrl>https://github.com/moq/Moq.AutoMocker</PackageProjectUrl>
     <Tags>moq;automocking;testing;TDD</Tags>
 
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <!-- Suppress CS1591: Missing XML comment for publicly visible type or member -->
-    <WarningsNotAsErrors>$(WarningsNotAsErrors),CS1591</WarningsNotAsErrors>
 
     <PublishRepositoryUrl>true</PublishRepositoryUrl>
     <DebugType>embedded</DebugType>
